+++ src/map/rayon.rs
@@ -126,7 +126,7 @@ where
     K: Send,
     V: Send,
 {
-    /// Sort the map’s key-value pairs in parallel, by the default ordering of the keys.
+    /// Sort the map's key-value pairs in parallel, by the default ordering of the keys.
     pub fn par_sort_keys(&mut self)
     where
         K: Ord,
@@ -134,7 +134,7 @@ where
         self.inner.par_sort_keys();
     }
 
-    /// Sort the map’s key-value pairs in place and in parallel, using the comparison
+    /// Sort the map's key-value pairs in place and in parallel, using the comparison
     /// function `cmp`.
     ///
     /// The comparison function receives two key and value pairs to compare (you
@@ -153,6 +153,16 @@ where
         F: Fn(&K, &V, &K, &V) -> Ordering + Sync,
     {
         self.inner.par_sorted_by(cmp)
+    }
+
+    /// Sort the map's key-value pairs in place and in parallel, using a sort-key extraction
+    /// function.
+    pub fn par_sort_by_key<T, F>(&mut self, sort_key: F)
+    where
+        T: Ord,
+        F: Fn(&K, &V) -> T + Sync,
+    {
+        self.inner.par_sort_by_key(sort_key)
     }
 
     /// Sort the map's key-value pairs in parallel, by the default ordering of the keys.
@@ -184,7 +194,17 @@ where
         self.inner.par_sorted_unstable_by(cmp)
     }
 
-    /// Sort the map’s key-value pairs in place and in parallel, using a sort-key extraction
+    /// Sort the map's key-value pairs in place and in parallel, using a sort-key extraction
+    /// function.
+    pub fn par_sort_unstable_by_key<T, F>(&mut self, sort_key: F)
+    where
+        T: Ord,
+        F: Fn(&K, &V) -> T + Sync,
+    {
+        self.inner.par_sort_unstable_by_key(sort_key)
+    }
+
+    /// Sort the map's key-value pairs in place and in parallel, using a sort-key extraction
     /// function.
     pub fn par_sort_by_cached_key<T, F>(&mut self, sort_key: F)
     where